Cleanup helper classes and move them to their own files
