+++ DynamicLinqPadPostgreSqlDriver/DynamicLinqPadPostgreSqlDriver.csproj
@@ -76,6 +76,8 @@
     <Compile Include="Extensions\TableDataExtensions.cs" />
     <Compile Include="Extensions\TypeBuilderExtensions.cs" />
     <Compile Include="Extensions\TypeExtensions.cs" />
+    <Compile Include="FunctionData.cs" />
+    <Compile Include="FunctionReturnTypeInfo.cs" />
     <Compile Include="IDatabaseObjectProvider.cs" />
     <Compile Include="PostgreSqlFunctionsProvider.cs" />
     <Compile Include="PostgreSqlTablesProvider.cs" />
