+++ DynamicLinqPadPostgreSqlDriver/FunctionData.cs
@@ -1,0 +1,13 @@
+namespace DynamicLinqPadPostgreSqlDriver
+{
+   internal class FunctionData
+   {
+      public string Name { get; set; }
+      public string ReturnType { get; set; }
+      public int ArgumentCount { get; set; }
+      public string[] ArgumentNames { get; set; }
+      public int[] ArgumentTypeOids { get; set; }
+      public object[] ArgumentDefaults { get; set; }
+      public bool IsMultiValueReturn { get; set; }
+   }
+}
