+++ DynamicLinqPadPostgreSqlDriver/FunctionReturnTypeInfo.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace DynamicLinqPadPostgreSqlDriver
+{
+   internal class FunctionReturnTypeInfo
+   {
+      public bool ExistsAsTable { get; set; }
+      public Type ElementType { get; set; }
+      public Type CollectionType { get; set; }
+   }
+}
